Wizard footer changes
