+++ extensions/azurePublishNew/src/components/provisioningWizards/footers/WizardFooter.tsx
@@ -4,9 +4,21 @@
 import * as React from 'react';
 import formatMessage from 'format-message';
 import styled from '@emotion/styled';
-import { DefaultButton } from 'office-ui-fabric-react';
+import { DefaultButton, PersonaSize } from 'office-ui-fabric-react';
 
+import { UserPersona } from '../../shared/userPersona/UserPersona';
 import { WizardStep } from '../../shared/wizard';
+import { UserInfo } from '../../../types';
+
+type ProvisonActionsStylingProps = {
+  showSignout: boolean;
+};
+
+const ProvisonActions = styled.div<ProvisonActionsStylingProps>((props) => ({
+  display: 'flex',
+  flexFlow: 'row nowrap',
+  justifyContent: props.showSignout ? 'space-between' : 'flex-end',
+}));
 
 const FooterButton = styled(DefaultButton)`
   margin-right: 8px;
@@ -18,29 +30,34 @@ const ButtonContainer = styled.div`
   }
 `;
 
-type Props = WizardStep;
+type Props = { userInfo?: UserInfo } & WizardStep;
 
 export const WizardFooter = (props: Props) => {
-  const { navigationState: navigation } = props;
+  const { userInfo, navigationState: navigation } = props;
 
   return (
-    <ButtonContainer>
-      <FooterButton
-        disabled={!navigation.canGoBack}
-        text={navigation.backText || formatMessage('Back')}
-        onClick={() => props.onBack()}
-      />
-      <FooterButton
-        primary
-        disabled={!navigation.canGoNext}
-        text={navigation.nextText || formatMessage('Next')}
-        onClick={() => props.onNext()}
-      />
-      <FooterButton
-        disabled={!navigation.canCancel}
-        text={navigation.cancelText || formatMessage('Cancel')}
-        onClick={() => props.onCancel()}
-      />
-    </ButtonContainer>
+    <ProvisonActions showSignout={!!userInfo}>
+      {userInfo ? (
+        <UserPersona secondaryText={formatMessage('Sign out')} size={PersonaSize.size40} text={userInfo?.name} />
+      ) : null}
+      <ButtonContainer>
+        <FooterButton
+          disabled={!navigation.canGoBack}
+          text={formatMessage(navigation.backText || 'Back')}
+          onClick={() => props.onBack()}
+        />
+        <FooterButton
+          primary
+          disabled={!navigation.canGoNext}
+          text={formatMessage(navigation.nextText || 'Next')}
+          onClick={() => props.onNext()}
+        />
+        <FooterButton
+          disabled={!navigation.canCancel}
+          text={formatMessage(navigation.cancelText || 'Cancel')}
+          onClick={() => props.onCancel()}
+        />
+      </ButtonContainer>
+    </ProvisonActions>
   );
 };
